+++ frontend/web/js/components/group-selector.ts
@@ -1,5 +1,6 @@
 import { ApiGroupInfoResponse, MembershipEdges } from "../core/types";
 import { ChoiceSelect } from "../utils/choice-select";
+import { emitEvent } from "../utils/event-bus";
 
 const template = document.createElement("template");
 template.innerHTML = String.raw`
@@ -83,6 +84,8 @@ export class GroupSelector extends HTMLElement {
       return;
     }
 
+    let membershipChanged = false;
+
     try {
       // For each sample, compute which groups to add/remove
       // A sample should be in ALL selected groups and in NO deselected groups
@@ -98,12 +101,14 @@ export class GroupSelector extends HTMLElement {
         if (toAdd.length > 0) {
           for (const gid of toAdd) {
             await this.addToGroup!(gid, [sampleId]);
+            membershipChanged = true;
           }
         }
 
         if (toRemove.length > 0) {
           for (const gid of toRemove) {
             await this.removeFromGroup!(gid, [sampleId]);
+            membershipChanged = true;
           }
         }
       }
@@ -113,6 +118,10 @@ export class GroupSelector extends HTMLElement {
     } catch (err) {
       console.error("Updating sample group memberships failed", err);
       this.dispatchEvent(new CustomEvent("apply:error", { detail: { error: err }, bubbles: true }));
+    } finally {
+      if (membershipChanged) {
+        emitEvent("samples:group-memberships-changed", { sampleIds });
+      }
     }
   };